requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
requests>=2.31.0
openrouteservice>=2.3.0
geopy>=2.4.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
# Streamlit - install separately if you want the web UI
# On Windows with Python 3.14, you may need Visual Studio Build Tools
# See STREAMLIT_SETUP.md for installation instructions
# streamlit>=1.28.0